refactor: Extract dic client
